Split texture tab in list and view
