+++ gapic/src/main/com/google/gapid/models/Analytics.java
@@ -34,7 +34,7 @@ public class Analytics implements ExceptionHandler {
     Main, LeftTabs, RightTabs,
     About, Help, GotoCommand, GotoMemory, Licenses, Settings, Trace, Welcome,
     // See MainWindow.MainTab.Type
-    FilmStrip, Profile, Commands, Framebuffer, Pipeline, Textures, Geometry, Shaders, Performance, Report, Log, State, Memory,
+    FilmStrip, Profile, Commands, Framebuffer, Pipeline, TextureList, TextureView, Geometry, Shaders, Performance, Report, Log, State, Memory,
     ContextSelector, ReplayDeviceSelector, QueryMetadata;
   }
